Create Sharjah International Conference on Linguistic Intelligence Announced by AUS and Emirates Scholar
Sharjah, May 23, 2024: Under the patronage of His Highness Sheikh Dr. Sultan bin Mohammed Al Qasimi, Member of the UAE Supreme Council and Ruler of Sharjah, the Emirate of Sharjah is set to host the Sharjah International Conference for Linguistic Intelligence, set to take place from October 17 to 19, 2024. Organized by American University of Sharjah (AUS) and the Emirates Scholar Center for Research and Studies,  the conference aims to explore the intersections of linguistics, artificial intelligence and cultural development. 

The Sharjah International Conference on Linguistic Intelligence will serve as a premier global forum for academics, researchers and industry professionals to discuss the impact of linguistic diversity and technological advancements on community development and cultural identity. The conference will feature over 600 participants, more than 100 presentations and 20 workshops.  

Her Excellency Sheikha Bodour bint Sultan Al Qasimi, President of American University of Sharjah, expressed her pride in Sharjah’s hosting and sponsorship of the Sharjah International Conference for Linguistic Intelligence. She highlighted the emirate’s rich history and vibrant culture and emphasized her confidence in the conference’s ability to illuminate the path for researchers and experts to envision the future of linguistic learning using artificial intelligence technologies to develop cultural identity. 

Emirates Scholar Research Center Signs MoU with Global Investment Promotion Committee to Advance Collaborative Research and Innovationhttps://www.emiratesscholar.com/emirates-scholar-research-center-signs-mou-with-global-investment-promotion-committee-to-advance-collaborative-research-and-innovation/

The announcement of the conference was made during a press conference held on the AUS campus, where a Memorandum of Understanding (MoU) was signed between AUS and the Emirates Scholar Center for Research and Studies. The agreement was signed by Dr. Tod Laursen, AUS Chancellor, and Dr. Fawaz Habbal, Director-General of the Emirates Scholar and Board Member, in the presence of Dr. Firas Habbal, President of the Emirates Scholar Center and Vice Chancellor of its Board of Trustees along with the university’s Board of Trustees and members of the center’s Board of Trustees. This partnership will facilitate joint research initiatives, conferences and academic activities, fostering a vibrant exchange of ideas and expertise. 

His Excellency Dr. Abdullah Bilhaif Al Nuaimi, Chairman of Sharjah Consultative Council , and Chancellor  of the Emirates Scholar Center for Research and Studies, emphasized the importance of this conference in providing an opportunity for the global scientific community to collaborate and innovate in areas related to human language and natural language processing technology, contributing to the development of innovative solutions and deeper understanding of the human mind and various cultures. 

“This collaboration with the Emirates Scholar Center for Research and Studies represents a significant step forward in our commitment to advancing research and shaping the future landscape of academia and beyond. By pooling our resources and expertise, we aim to create impactful solutions that benefit not only our institutions but also the broader community. We are particularly excited about the potential this partnership holds for supporting vital policy areas and enhancing our network with both local and international partners,” said Dr. Laursen.  

Dr. Firas Habbal, President of the Emirates Scholar Center, and Vice Chairperson of its Board of Trustees, stated that the conference plays a significant role in developing policies and strategies that support progress in the fields of artificial and linguistic intelligence, as well as in promoting education and learning in these areas. He expressed his confidence in the strong scientific footprint this conference will leave in linguistic research. 

Key objectives of the Sharjah International Conference on Linguistic Intelligence include promoting global dialogue on linguistic diversity, exploring the impact of AI and natural language processing on linguistic studies and advocating for policies that support linguistic diversity as a cornerstone of cultural wealth. The event will also feature dynamic sessions on AI-driven language learning tools, language preservation and multilingual communication platforms. 

PRESSR: Sharjah International Conference on Linguistic Intelligence announced by AUS and Emirates Scholarhttps://www.tradingview.com/news/reuters.com,2024-05-23:newsml_Zaw817wyr:0-pressr-sharjah-international-conference-on-linguistic-intelligence-announced-by-aus-and-emirates-scholar/