Create Emirates scholar congratulates HE Dr. Abdullah Belhaif on receiving an honorary professorship.
The Emirates Scholar Center for Research and Studies extends its warmest congratulations and felicitations to His Excellency Dr. Abdullah Belhaif Al Nuaimi for being awarded an honorary professorship from the College of Energy and Science. 

As The Heriot-Watt University in Dubai awarded His Excellency Dr. Abdullah Belhaif Al Nuaimi, Chairperson of the Sharjah Advisory Council and Chairperson of the Board of Trustees of the Emirates Scholar Center for Research and Studies on the honor of Professorship in Energy and Geological Sciences and Infrastructure, in recognition of his outstanding contribution to environmental, climate, and sustainability issues. 

Professor Heather McGregor, Dean and Deputy Vice Chancellor of the University, expressed pride in the university’s honoring of His Excellency, for his multiple achievements in energy, sciences, infrastructure, sustainability, climate change, community development, academic and government services

. Emirates Scholar Center for Research and Studies welcomed by His Excellency Dr. Abdullah Belhaif Al Nuaimi.https://www.emiratesscholar.com/emirates-scholar-center-for-research-and-studies-welcomed-by-his-excellency-dr-abdullah-belhaif-al-nuaimi/

Dr. Raja Issa Al Gurg, President of Heriot-Watt University Dubai, and Chairman of the Board and CEO of Easa Saleh Al Gurg Group (ESAG), presented the academic degrees. 

 She expressed her happiness at the recognition of His Excellency Dr. Abdullah Belhaif for his overall achievements and efforts towards community development, noting that H.E serves as an exceptional example capable of making a meaningful impact on society, inspiring future Emirati generations. 

On the other side, Dr. Abdullah Belhaif Al Nuaimi expressed his gratitude and appreciation for receiving the honorary professorship from Heriot-Watt University in Dubai, adding that obtaining this degree is recognition of his professional career and achievements in various fields. 

He expressed feeling deeply humbled and honored by this prestigious recognition, which he considers an honor for himself and the field in which he works. He promised everyone to fulfill the responsibility entrusted to him, exerting his utmost effort to contribute to the advancement and development of the fields of energy, sciences, environment, and sustainability. 

His Excellency Dr. Abdullah Belhaif Al Nuaimi is a prominent figure in both the public and academic spheres in the United Arab Emirates. He has held distinguished ministerial positions and membership in the Cabinet of the United Arab Emirates, most notably in the Ministry of Climate Change and Environment and the Ministry of Infrastructure Development.  

Through these positions, he led sustainable development and local and global climate change initiatives and spearheaded infrastructure projects. His tenure witnessed the launch of countless initiatives, such as research laboratories and the Sheikh Zayed Housing Program. He played a pivotal role in the Federal Authority for Land and Maritime Transport. 

In the academic field, His Excellency collaborated with Heriot-Watt University Dubai, where he chaired the Board of Directors of the Center of Excellence for Smart and Sustainable Construction (CESC) and shared knowledge as a visiting professor at the Dubai Higher Colleges of Technology. 

Emirati Accomplishments Celebrated with Honorary Degrees at Heriot-Watt University Dubai’s Graduation Ceremonieshttps://www.hw.ac.uk/news/articles/2023/emirati-accomplishments-celebrated-with.htm